added stack and queue
now, we have stacks and queues. a deque will come soon.

the stack is implemented by the list_st (pushing to front/poping from front).

the queue is implemented by stack. (infact we just bend it to push to back and
still pop from front). So if you read some errors with stacks, when working with
queues it is correct. ;)